refactor dca client
